Assignment-8/requirements.txt
# Required libraries

python-dotenv
requests
beautifulsoup4
pandas
openpyxl
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib

